De-duplicate NullString
