+++ payment/order.go
@@ -1,62 +1,41 @@
 package payment
 
 import (
-	"database/sql"
-	"encoding/json"
 	"strconv"
 	"strings"
 	"time"
 
+	"github.com/brave-intl/bat-go/utils/datastore"
 	uuid "github.com/satori/go.uuid"
 	"github.com/shopspring/decimal"
 	macaroon "gopkg.in/macaroon.v2"
 )
 
-// NullString is a type that lets ya get a null field from the database
-type NullString struct {
-	sql.NullString
-}
-
-// MarshalJSON for NullString
-func (ns *NullString) MarshalJSON() ([]byte, error) {
-	if !ns.Valid {
-		return []byte("null"), nil
-	}
-	return json.Marshal(ns.String)
-}
-
-// UnmarshalJSON unmarshalls NullString
-func (ns *NullString) UnmarshalJSON(data []byte) error {
-	ns.String = strings.Trim(string(data), `"`)
-	ns.Valid = true
-	return nil
-}
-
 // Order includes information about a particular order
 type Order struct {
-	ID         uuid.UUID       `json:"id" db:"id"`
-	CreatedAt  time.Time       `json:"createdAt" db:"created_at"`
-	Currency   string          `json:"currency" db:"currency"`
-	UpdatedAt  time.Time       `json:"updatedAt" db:"updated_at"`
-	TotalPrice decimal.Decimal `json:"totalPrice" db:"total_price"`
-	MerchantID string          `json:"merchantId" db:"merchant_id"`
-	Location   NullString      `json:"location" db:"location"`
-	Status     string          `json:"status" db:"status"`
-	Items      []OrderItem     `json:"items"`
+	ID         uuid.UUID            `json:"id" db:"id"`
+	CreatedAt  time.Time            `json:"createdAt" db:"created_at"`
+	Currency   string               `json:"currency" db:"currency"`
+	UpdatedAt  time.Time            `json:"updatedAt" db:"updated_at"`
+	TotalPrice decimal.Decimal      `json:"totalPrice" db:"total_price"`
+	MerchantID string               `json:"merchantId" db:"merchant_id"`
+	Location   datastore.NullString `json:"location" db:"location"`
+	Status     string               `json:"status" db:"status"`
+	Items      []OrderItem          `json:"items"`
 }
 
 // OrderItem includes information about a particular order item
 type OrderItem struct {
-	ID          uuid.UUID       `json:"id" db:"id"`
-	OrderID     uuid.UUID       `json:"orderId" db:"order_id"`
-	CreatedAt   *time.Time      `json:"createdAt" db:"created_at"`
-	UpdatedAt   *time.Time      `json:"updatedAt" db:"updated_at"`
-	Currency    string          `json:"currency" db:"currency"`
-	Quantity    int             `json:"quantity" db:"quantity"`
-	Price       decimal.Decimal `json:"price" db:"price"`
-	Subtotal    decimal.Decimal `json:"subtotal"`
-	Location    NullString      `json:"location" db:"location"`
-	Description NullString      `json:"description" db:"description"`
+	ID          uuid.UUID            `json:"id" db:"id"`
+	OrderID     uuid.UUID            `json:"orderId" db:"order_id"`
+	CreatedAt   *time.Time           `json:"createdAt" db:"created_at"`
+	UpdatedAt   *time.Time           `json:"updatedAt" db:"updated_at"`
+	Currency    string               `json:"currency" db:"currency"`
+	Quantity    int                  `json:"quantity" db:"quantity"`
+	Price       decimal.Decimal      `json:"price" db:"price"`
+	Subtotal    decimal.Decimal      `json:"subtotal"`
+	Location    datastore.NullString `json:"location" db:"location"`
+	Description datastore.NullString `json:"description" db:"description"`
 }
 
 // CreateOrderItemFromMacaroon creates an order item from a macaroon
